requirements.txt
Django==4.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.0
